Decorate sequential links, add sequence metadata
Even though sequential links already exist in some pages, we want to make sure they are decorated with the proper classes in so on (in the future, we could or otherwise transform the markup, too). This keeps the markdown focused on content and leaves the styling and so on to the `developers` repo.